basics

+++ 02_creating_an_empty_image.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 height = 300
 width = 400 
 
-image = np.zeros((height, width, 3)) 
+image = np.zeros((height, width, 3), dtype=np.uint8) 
 
 plt.imshow(image,cmap='gray') # display the image using matplotlib 
 plt.axis("off") # turn off axis labels
